refactor: modify components file export
