+++ frontend/src/pages/sign_up.jsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 import { Form, Modal, Button, DatePicker } from 'antd';
 import { css } from '@emotion/react';
 import useInput from '../hooks/useInput';
-import AppLayout from '../components/AppLayout';
+import { AppLayout } from '../components';
 import axios from 'axios';
 import Router, { useRouter } from 'next/router';
